增加Command结构,思考初始化流程和方法，处理command解析 优化整理file的socket发送 完成master的上传和下载命令处理，file的socket发送 upload、download task clear 完成大部分初始化、管理命令处理 下载任务超时计数、上传下载处理 完成下载Store命令处理 store server上传命令处理 下一步： backup错处理。 上传MD5验证计算 其他命令处理：index、delete

+++ config/src/main/java/com/ly/rhdfs/config/ServerConfig.java
@@ -11,7 +11,7 @@ import com.ly.common.constant.ParamConstants;
 public class ServerConfig {
 
     @Value("${server.id}")
-    private int currentServerId;
+    private long currentServerId;
     @Value("${server.port}")
     private int port;
     @Value("${master.server.config.path}")
@@ -67,11 +67,11 @@ public class ServerConfig {
     public void setTmpFileSuffix(String tmpFileSuffix) {
         this.tmpFileSuffix = tmpFileSuffix;
     }
-    public int getCurrentServerId() {
+    public long getCurrentServerId() {
         return currentServerId;
     }
 
-    public void setCurrentServerId(int currentServerId) {
+    public void setCurrentServerId(long currentServerId) {
         this.currentServerId = currentServerId;
     }
